+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Version 2.0.12 *(2026-<M>-<D>)*
 * Implement `__mod__` for builtin float.
 * Fix Codegen for select in GPU Compiler.
 * Implement `__mul__` for builtin list.
+* Extend ListComp, SetComp and DictComp to support targets, eg `[x + y for x, y in ((1, 2), (2, 3), (4, 5))]`.
 
 Version 2.0.11 *(2026-07-01)*
 -----------------------------